move from `_bstr.py` to `comtypes/_post_coinit/bstr.py`
