+++ comtypes/_post_coinit/bstr.py
